Bug Fix for callback function to check callable and accommodate 4 parameters
Bug Fix: Accommodates class methods, which will have an additional self parameter, making a total of 4 parameters in addition to 3. 
Bug Fix: Use the callable() built-in function to check if callback is a callable object, as it will handle all types of functions, methods, and any object implementing the __call__ method.